indian_banks/requirements.txt
Django==2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
pytz==2019.3
